week08/p17_webLLM/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langgraph==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
dashscope
psycopg2-binary>=2.9.0
gradio>=4.0.0
pandas>=2.0.0
requests>=2.31.0
celery==5.2.7
redis>=5.0.0
flower>=2.0.0